Rename class, variable, property and move another folder
