+++ app/src/main/java/team/triplog/presentation/main/fragment/MainTripPlanFragment.kt
@@ -10,7 +10,7 @@ import androidx.lifecycle.Observer
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import team.triplog.databinding.FragmentMainTripPlanBinding
 import team.triplog.domain.model.Trip
-import team.triplog.presentation.activity.startTripInfoActivity
+import team.triplog.presentation.trip.add.activity.startTripInfoActivity
 import team.triplog.presentation.main.adapter.MainTripPlanAdapter
 import team.triplog.presentation.main.viewmodel.MainTripPlanViewModel
