+++ app/src/main/java/team/triplog/presentation/trip/add/adapter/TripCheckAdapter.kt
@@ -1,4 +1,4 @@
-package team.triplog.presentation.adapter
+package team.triplog.presentation.trip.add.adapter
 
 import android.content.Context
 import android.view.LayoutInflater
