+++ app/src/main/res/values/colors.xml
@@ -29,10 +29,10 @@
     <color name="main_more_version">@color/gray_999999</color>
     <color name="more_user_title">@color/black_000000</color>
     <color name="more_user_contents">@color/gray_666666</color>
-    <color name="trip_info_description">@color/gray_333333</color>
-    <color name="trip_info_hint">@color/gray_999999</color>
-    <color name="trip_info_trip_name">@color/main</color>
-    <color name="trip_info_select_date">@color/gray_333333</color>
+    <color name="trip_add_description">@color/gray_333333</color>
+    <color name="trip_add_hint">@color/gray_999999</color>
+    <color name="trip_add_trip_name">@color/main</color>
+    <color name="trip_add_select_date">@color/gray_333333</color>
     <color name="floating_button">@color/main</color>
     <color name="trip_check_item">@color/black_000000</color>
     <color name="trip_check_pic">@color/gray_999999</color>
