+++ app/src/main/res/values/strings.xml
@@ -31,10 +31,10 @@
     <string name="main_more_setting_notice_advertising_desc">알림을 받으려면 활성화 해주세요.</string>
     <string name="main_trip_log_title">여행 기록</string>
     <string name="main_trip_log_save">저장</string>
-    <string name="trip_info_description">설레는 여행을 시작할 준비가 되셨나요?\n준비가 되었다면 여행의 이름을 지어주세요.</string>
-    <string name="trip_info_name_hint">여행 이름</string>
-    <string name="trip_info_name_limit">* 최대 20자까지 입력 가능합니다.</string>
-    <string name="trip_info_select_date">여행의 기간도 설정해주세요.</string>
+    <string name="trip_add_description">설레는 여행을 시작할 준비가 되셨나요?\n준비가 되었다면 여행의 이름을 지어주세요.</string>
+    <string name="trip_add_name_hint">여행 이름</string>
+    <string name="trip_add_name_limit">* 최대 20자까지 입력 가능합니다.</string>
+    <string name="trip_add_select_date">여행의 기간도 설정해주세요.</string>
     <string name="trip_check_title">준비물</string>
     <string name="trip_check_pic">담당자 : %s</string>
     <string name="trip_member_title">일행</string>
